+++ pkg/gen/input/apptest/internal/params/key.go
@@ -1,0 +1,13 @@
+package params
+
+func AppName(p Params) string {
+	return p.AppName
+}
+
+func RepoName(p Params) string {
+	return p.RepoName
+}
+
+func Catalog(p Params) string {
+	return p.Catalog
+}