+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added a new `ci` command with a `setup` subcommand that configures webhooks to our Tekton installation
+- Added a new `gen apptest` command that creates the files needed by apptest-framework
 
 ## [6.21.0] - 2024-03-07